refactor(grammar): hide grammar
